instruments.py:
added function read_resistance for the Keithley2600. This function first finds the proper range for the device and then returns the resistance of the device

Keithley_tek_bias_tee.py:

added new measurement routine, which determines the proper resistance before and after the pulse. designed for pcm devices and viktors setup